Removed ZEN related bug
There is an issue if ZEN and ZEI are loaded at the same time and the interface is opened using the "openCuratorInterface" command.
ZEI would mix up all custom modules added per script from ZEN.
The only way to fix it was to manually get out of Zeus and back in.
However, this commit should fix the issue by adding native ZEN modules when necessary.
This bug did not occur if the client was the server (host).