Feature: level of detail for interior fill
Added a UI option to constrain the "level of detail" of interior fill.

This allows the user to only place defences (i.e. sandbags, bunkers, and shoothouse walls) in order to minimise the number of objects placed in a mission.

Currently could be replaced with a checkbox, but a dropdown allows for other premade detail-level "templates" in future